move dandi to optional; reorganize tool imports; make dandi import optional
