+++ requirements-minimal.txt
@@ -6,8 +6,9 @@ h5py>=3.9.0
 hdmf>=3.11.0
 hdmf_zarr>=0.4.0
 pynwb>=2.3.2;python_version>='3.8'
+nwbinspector>=0.4.31
+pydantic>=1.10.13,<2.0.0
 psutil>=5.8.0
 tqdm>=4.60.0
-dandi>=0.57.0
 pandas
 parse>=1.20.0
