+++ src/neuroconv/tools/__init__.py
@@ -1,3 +1,4 @@
+"""Collection of all helper functions that require at least one external dependency (some being optional as well)."""
 from .importing import get_package
 from .nwb_helpers import get_module
 from .path_expansion import LocalPathExpander
