+++ src/neuroconv/tools/data_transfers/_globus.py
@@ -1,29 +1,17 @@
 """Collection of helper functions for assessing and performing automated data transfers."""
 import json
-import os
 import re
 from pathlib import Path
-from shutil import rmtree
-from tempfile import mkdtemp
 from time import sleep, time
-from typing import Dict, List, Optional, Tuple, Union
-from warnings import warn
+from typing import Dict, List, Tuple, Union
 
-from dandi.download import download as dandi_download
-from dandi.organize import organize as dandi_organize
-from dandi.upload import upload as dandi_upload
-from pynwb import NWBHDF5IO
+from nwbinspector.utils import is_module_installed
+from pydantic import DirectoryPath
 from tqdm import tqdm
 
 from .processes import deploy_process
-from ..utils import FolderPathType, OptionalFolderPathType
 
-try:  # pragma: no cover
-    import globus_cli
-
-    HAVE_GLOBUS = True
-except ModuleNotFoundError:
-    HAVE_GLOBUS = False
+HAVE_GLOBUS = is_module_installed(module_name="globus_cli")
 
 
 def get_globus_dataset_content_sizes(
@@ -50,7 +38,7 @@ def transfer_globus_content(
     source_endpoint_id: str,
     source_files: Union[str, List[List[str]]],
     destination_endpoint_id: str,
-    destination_folder: FolderPathType,
+    destination_folder: DirectoryPath,
     display_progress: bool = True,
     progress_update_rate: float = 60.0,
     progress_update_timeout: float = 600.0,
@@ -203,167 +191,3 @@ def transfer_globus_content(
         progress_update_timeout=progress_update_timeout,
     )
     return success, list(task_total_sizes)
-
-
-def estimate_total_conversion_runtime(
-    total_mb: float,
-    transfer_rate_mb: float = 20.0,
-    conversion_rate_mb: float = 17.0,
-    upload_rate_mb: float = 40,
-    compression_ratio: float = 1.7,
-):
-    """
-    Estimate how long the combined process of data transfer, conversion, and upload is expected to take.
-
-    Parameters
-    ----------
-    total_mb: float
-        The total amount of data (in MB) that will be transferred, converted, and uploaded to dandi.
-    transfer_rate_mb : float, default: 20.0
-        Estimate of the transfer rate for the data.
-    conversion_rate_mb : float, default: 17.0
-        Estimate of the conversion rate for the data. Can vary widely depending on conversion options and type of data.
-        Figure of 17MB/s is based on extensive compression of high-volume, high-resolution ecephys.
-    upload_rate_mb : float, default: 40.0
-        Estimate of the upload rate of a single file to the DANDI archive.
-    compression_ratio : float, default: 1.7
-        Estimate of the final average compression ratio for datasets in the file. Can vary widely.
-    """
-    c = 1 / compression_ratio  # compressed_size = total_size * c
-    return total_mb * (1 / transfer_rate_mb + 1 / conversion_rate_mb + c / upload_rate_mb)
-
-
-def estimate_s3_conversion_cost(
-    total_mb: float,
-    transfer_rate_mb: float = 20.0,
-    conversion_rate_mb: float = 17.0,
-    upload_rate_mb: float = 40.0,
-    compression_ratio: float = 1.7,
-):
-    """
-    Estimate potential cost of performing an entire conversion on S3 using full automation.
-
-    Parameters
-    ----------
-    total_mb: float
-        The total amount of data (in MB) that will be transferred, converted, and uploaded to dandi.
-    transfer_rate_mb : float, default: 20.0
-        Estimate of the transfer rate for the data.
-    conversion_rate_mb : float, default: 17.0
-        Estimate of the conversion rate for the data. Can vary widely depending on conversion options and type of data.
-        Figure of 17MB/s is based on extensive compression of high-volume, high-resolution ecephys.
-    upload_rate_mb : float, default: 40.0
-        Estimate of the upload rate of a single file to the DANDI Archive.
-    compression_ratio : float, default: 1.7
-        Estimate of the final average compression ratio for datasets in the file. Can vary widely.
-    """
-    c = 1 / compression_ratio  # compressed_size = total_size * c
-    total_mb_s = (
-        total_mb**2 / 2 * (1 / transfer_rate_mb + (2 * c + 1) / conversion_rate_mb + 2 * c**2 / upload_rate_mb)
-    )
-    cost_gb_m = 0.08 / 1e3  # $0.08 / GB Month
-    cost_mb_s = cost_gb_m / (1e3 * 2.628e6)  # assuming 30 day month; unsure how amazon weights shorter months?
-    return cost_mb_s * total_mb_s
-
-
-def automatic_dandi_upload(
-    dandiset_id: str,
-    nwb_folder_path: FolderPathType,
-    dandiset_folder_path: OptionalFolderPathType = None,
-    version: str = "draft",
-    staging: bool = False,
-    cleanup: bool = False,
-    number_of_jobs: Optional[int] = None,
-    number_of_threads: Optional[int] = None,
-):
-    """
-    Fully automated upload of NWBFiles to a DANDISet.
-
-    Requires an API token set as an envrinment variable named DANDI_API_KEY.
-
-    To set this in your bash terminal in Linux or MacOS, run
-        export DANDI_API_KEY=...
-    or in Windows
-        set DANDI_API_KEY=...
-
-    DO NOT STORE THIS IN ANY PUBLICLY SHARED CODE.
-
-    Parameters
-    ----------
-    dandiset_id : str
-        Six-digit string identifier for the DANDISet the NWBFiles will be uploaded to.
-    nwb_folder_path : folder path
-        Folder containing the NWBFiles to be uploaded.
-    dandiset_folder_path : folder path, optional
-        A separate folder location within which to download the dandiset.
-        Used in cases where you do not have write permissions for the parent of the 'nwb_folder_path' directory.
-        Default behavior downloads the DANDISet to a folder adjacent to the 'nwb_folder_path'.
-    version : {None, "draft", "version"}
-        The default is "draft".
-    staging : bool, default: False
-        Is the DANDISet hosted on the staging server? This is mostly for testing purposes.
-        The default is False.
-    cleanup : bool, default: False
-        Whether to remove the dandiset folder path and nwb_folder_path.
-        Defaults to False.
-    number_of_jobs : int, optional
-        The number of jobs to use in the DANDI upload process.
-    number_of_threads : int, optional
-        The number of threads to use in the DANDI upload process.
-    """
-    assert os.getenv("DANDI_API_KEY"), (
-        "Unable to find environment variable 'DANDI_API_KEY'. "
-        "Please retrieve your token from DANDI and set this environment variable."
-    )
-
-    dandiset_folder_path = (
-        Path(mkdtemp(dir=nwb_folder_path.parent)) if dandiset_folder_path is None else dandiset_folder_path
-    )
-    dandiset_path = dandiset_folder_path / dandiset_id
-    # Odd big of logic upstream: https://github.com/dandi/dandi-cli/blob/master/dandi/cli/cmd_upload.py#L92-L96
-    if number_of_threads is not None and number_of_threads > 1 and number_of_jobs is None:
-        number_of_jobs = -1
-
-    url_base = "https://gui-staging.dandiarchive.org" if staging else "https://dandiarchive.org"
-    dandiset_url = f"{url_base}/dandiset/{dandiset_id}/{version}"
-    dandi_download(urls=dandiset_url, output_dir=str(dandiset_folder_path), get_metadata=True, get_assets=False)
-    assert dandiset_path.exists(), "DANDI download failed!"
-
-    # TODO: need PR on DANDI to expose number of jobs
-    dandi_organize(
-        paths=str(nwb_folder_path), dandiset_path=str(dandiset_path), devel_debug=True if number_of_jobs == 1 else False
-    )
-    organized_nwbfiles = dandiset_path.rglob("*.nwb")
-
-    # DANDI has yet to implement forcing of session_id inclusion in organize step
-    # This manually enforces it when only a single session per subject is organized
-    for organized_nwbfile in organized_nwbfiles:
-        if "ses" not in organized_nwbfile.stem:
-            with NWBHDF5IO(path=organized_nwbfile, mode="r") as io:
-                nwbfile = io.read()
-                session_id = nwbfile.session_id
-            dandi_stem = organized_nwbfile.stem
-            dandi_stem_split = dandi_stem.split("_")
-            dandi_stem_split.insert(1, f"ses-{session_id}")
-            corrected_name = "_".join(dandi_stem_split) + ".nwb"
-            organized_nwbfile.rename(organized_nwbfile.parent / corrected_name)
-    organized_nwbfiles = dandiset_path.rglob("*.nwb")
-    # The above block can be removed once they add the feature
-
-    assert len(list(dandiset_path.iterdir())) > 1, "DANDI organize failed!"
-
-    dandi_instance = "dandi-staging" if staging else "dandi"  # Test
-    dandi_upload(
-        paths=[str(x) for x in organized_nwbfiles],
-        dandi_instance=dandi_instance,
-        jobs=number_of_jobs,
-        jobs_per_file=number_of_threads,
-    )
-
-    # Cleanup should be confirmed manually; Windows especially can complain
-    if cleanup:
-        try:
-            rmtree(path=dandiset_folder_path)
-            rmtree(path=nwb_folder_path)
-        except PermissionError:  # pragma: no cover
-            warn("Unable to clean up source files and dandiset! Please manually delete them.")
